+++ lib/Docstring.mli
@@ -20,12 +20,6 @@ val parse_file : Lexing.position -> string -> Odoc_parser.Ast.t
 
 val warn : Format.formatter -> Odoc_parser.Warning.t -> unit
 
-type error =
-  | Moved of Location.t * Location.t * string
-  | Unstable of Location.t * string * string
-  | Added of Location.t * string
-  | Removed of Location.t * string
-
 val is_tag_only : Odoc_parser.Ast.t -> bool
 (** [true] if the documentation only contains tags *)