+++ CHANGES.md
@@ -20,6 +20,8 @@
 
 ### New features
 
+- Improved error reporting for unstable or dropped comments (#2292, @gpetiot)
+
 ## 0.25.1 (2023-03-06)
 
 ### Bug fixes